feat: rebrand to Agent Chronicle with st.navigation, dark theme, sidebar footer
- Restructure app.py as pure st.navigation entry point for explicit sidebar labels
- Create pages/0_Overview.py with dashboard charts (moved from app.py)
- Add branding.py with shared sidebar header, footer, and CSS
- Add .streamlit/config.toml with dark slate theme (green primary)
- Fix tertiary button red hover color to gray via CSS override
- Remove set_page_config from page files (single call in app.py)
- Sidebar now shows Overview/Session Browser/SQL Query (no more 'app')
