+++ examples/explorer/.streamlit/config.toml
@@ -1,0 +1,6 @@
+[theme]
+primaryColor = "#a3e635"
+backgroundColor = "#0f172a"
+secondaryBackgroundColor = "#1e293b"
+textColor = "#e2e8f0"
+font = "sans serif"
