+++ examples/explorer/pages/2_SQL_Query.py
@@ -10,7 +10,6 @@ import sys, os
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from db import get_connection, get_data_paths, run_query
 
-st.set_page_config(page_title="SQL Query", page_icon="🔎", layout="wide")
 st.title("🔎 SQL Query")
 
 con = get_connection()
